Fix issue with how custom page filtering is layered in
This moves the API identifiers into the artemisapi library so that there
is not an import issue when this is used in the lambda environment
(where not all API packages are going to be present at the same time).
