--- backend/lambdas/api/search_repositories/search_repositories/util/const.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-
-class SearchRepositoriesAPIIdentifier(Enum):
-    GET = "search_repositories_get"
